Variável com (float)

+++ exercio1.php
@@ -1,7 +1,7 @@
 <?php
 
 print "Valor em metros: \n";
-$valormetros = fgets (STDIN);
+$valormetros = (float)fgets (STDIN);
 
 $valormilimetros= $valormetros*1000;
